Added Submit Button script and demo
When added as an Action to an Acrobat PDF form, this script scans for empty mandatory fields and alerts the user if the form is incomplete. If the form is complete, an email is generated with the PDF as an attachment and, optionally, with data pulled from specified form fields.